filemon: add enabled toggle
When the filemon buffer is read its entries are lost forever, so if
someone inadvertently cat's it for example, without being prepared to
properl consume the output we will lose potentially useful data.

To prevent such accidental reading from occuring add a new toggle
which has to be set to one to enable consuming the buffer. Upon
draining the buffer this toggle is automatically reset back to zero.

While at it also use the occasion to rename some of the internal
structures