filemon: Implement view size limit
Implement a way of limiting the amount of entries shown when the
change buffer is queried. This is necessary since on very busy servers
it is possible that a client consuming the buffer might spend forever,
constantly showing 1 entry at a time.

To set the number of entries displayed when querying /proc/filemon/buffer,
one should set /proc/filemon/listing_limit