filemon: Implement directory listing filtering
Sometimes we might want to limit the directories which are monitored
for changes. To this effect extend the procfs interface to be able to
add directories which act as a filter for events. When an entry is added
to the "path_filter" it is added to a list. Later, whenever a VFS event
occurs the dentry of the modified file is checked to see if it is a child
of any of the added filter paths

This patch contains the necessary infrastructure to be able to add/list
directories which we want to filter.