filemon: add the ability to delete filter path
The way this works is by assgining a global, incremental id to each
filter path that is being added. When the "path_filter" proc file is
cat'ed it will list this id in square brackets e.g. [0]: /path . To
delete the filter "/path" just do an "echo 0 > /proc/filemon/path_filter_delete"