filemon: Fix resource leak bug
It was possible to experience a resource leak if after having called
the filemon_seq_start, filemon_seq_stop is called, since resources were
freed in filemon_seq_next. Refactor the code so that we call the
necessary release logic in filemon_seq_stop. It is now possible to
umount a file system correctly if listing has been interrupted.