filemon: Fix handling of VERY long path names
Stress testing the the filemon code exposed an issue when handling
paths which are longer than the statically defined buffer of 4k.
This failure mode can lead to machine starvation as has been
observed during live testing.

To cope with situation like these change the buffer to be dynamically
allocated and also expose a user-space proc file which allows one to
set the maximum size of the path buffer. In cases when the buffer's
size is insufficient to store the full path just drop it and produce
a warning, giving a chance to the sys admin to increase the buffer
further. Also, change the default size of the buffer to be 8192
bytes.

(cherry picked from commit 97b4d7bd7529a1b329e3f291d5eca732a9f27d63)

Conflicts:
	fs/notify/filemon.c